test_drive/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
matplotlib>=3.3.2
scikit-learn>=0.24.2
seaborn>=0.11.1
Pillow>=8.0.1
PyWavelets>=1.1.1
tensorboard>=2.4.0 